moved resource classes to dialogue.py
